Corrected multi-domain bug + added also-notify support
Yes, this is a large amount of changes. It consists of adding support
for also-notify, correcting an issue where puppet would fail when
you're managing multiple domains on a single IP and brought coding up
to puppetlabs standards.

Unfortunately, implementation of views would take a lot more work and
I've decided not to bother right now. =p