Added support for forward zone type
Modified the zone erb template to have the correct behaviour when
zone_type = forward.

a simple example zone is:
dns::zone { 'example.net':
  zone_type => 'forward',
  forwarders => ['10.0.0.1','10.0.0.2']
}

there's more work that could be done to make the solution more elegant,
but at least it works. the db file is still created, even though it is
not used.