MX preference fix, unique alias, add tests
Fixes ajjahn/puppet-dns#37
 - adds all fields to alias: host, zone, pref, and data
 - any three of the four can be the same for multiple MX records
   under management, which would previously cause a duplicate alias
 - add test to confirm concat fragment is named as expected

Fixes ajjahn/puppet-dns#39 by writing the promised tests

Other changes of note:
Default preference is now 10. A pref of 0 is allowed, but if a
DNS administrator wanted to insert a more-preferable MX record,
she would be unable to do so until the previous record TTL expired.

Default host is now '@'. This allows for cleaner manifests because
the most common MX case is to create an exchanger for the entire
$ORIGIN. BIND will do the substitution automatically.

Blank labels are permitted in zone files, but are not handled by
this module because it does not render concat fragments in an
explicit order. Since BIND will substitute the last valid label
for any blank label, the resulting substitution would be
unpredictable.

TTL test is not specific to MX records. It should validate all BIND
records, so is better placed in the dns::record define. Valid TTL
range is 0-2**31-1 according to Section 8 of RFC 2181. BIND also
accepts units of m (minutes), h (hours), d (days), and w (weeks),
so 1d4h is a valid TTL, too.