Add `all_ptr` parameter to `dns::record::a` record type
This adds the `all_ptr=>`_[true|false]_ parameter to the `dns::record::a` type and sets up a private defined type that allows the `dns::record::a` type to create multiple `dns::record::ptr` resources for a records with multiple `data` values.

This also creates a set of rspec tests for the `dns::record::a` defined type.