Add `all` and `first` values to `ptr` parameter to `dns::record::a` record type
This updates the `ptr` parameter of the `dns::record::a` type to allow `first` or `all` values.  `first` is the same as `true` (or any other `str2bool` true values) and will act as the traditional `ptr` parameter did.  With `all`, a `dns::record::ptr` resource is created for all of the ip addresses in `data`, not just the first.  This uses a private defined type `dns_reverse_ptr_record` to actually create the `dns::record::ptr` resources so that we can get iteration in the older puppet parsers.

This also creates a set of rspec tests for the `dns::record::a` defined type.