Add `serial_style` parameter to dns::zone
This adds an option to dns::zone to generate the serial number using
the default 'timestamp' style, where the serial number is the unix
timestamp at the time of serial generation, or a new 'date' style
that is the traditional YYYYMMDDXX serial number of zone files managed
by hand.

The serial number generation itself is simple, but the command to
apply the serial number to the zone file is now much more complex,
requiring an examination of the existing zone file to determine the
current serial number so that the new serial number can be incremented
above the current if needed.